everything works now, yay!

+++ exercise.wwwapi/DataModels/StudentTypes/PostStudent.cs
@@ -6,6 +6,6 @@ public class PostStudent
     public string LastName { get; set; }
     public DateTime DateOfBirth { get; set; }
     public int CourseId { get; set; }
-    public int AverageGrade { get; set; }
+    public double AverageGrade { get; set; }
     public DateTime StartDate { get; set; }
 }
